fixed up world collisson checking, added a colMatrix to forms so 2 solid forms can pass through eachother, updated Cell to accomodate multiple solid formms, and probly more sstuff ,this was a big one for Jumpo

+++ formglfw/WorldView.h
@@ -4,8 +4,8 @@ typedef struct WorldView {
 	float frameX;
 	float frameY;
 	float frame;
-	int centerX;
-	int centerY;
+	float centerX;
+	float centerY;
 	int scalePower;
 	//int defaultFrameX;
 	//int defaultFrameY;
